refactor: isolate reference ingestion from template briefs
