+++ apps/app/src/react-app/domains/session/templates/template-brief.ts
@@ -1,22 +1,11 @@
 import type { TemplateCategory, TemplateManifestV1 } from "@ipollowork/types/templates";
-import type { ComposerAttachment } from "@/app/types";
 import { t } from "@/i18n";
-import {
-  ingestReferenceFile,
-  isReferenceFile,
-  prepareOriginalReferenceAttachment,
-} from "@/react-app/domains/session/references/ingestion";
-import {
-  inferTemplateBriefFromIngestions,
-} from "@/react-app/domains/session/references/brief-autofill";
 
 export type TemplateBrief = {
   title: string;
   audience: string;
   details: string;
 };
-
-export const TEMPLATE_BRIEF_REFERENCE_MAX_BYTES = 25 * 1024 * 1024;
 
 export const TEMPLATE_BRIEF_REFERENCE_ACCEPT = [
   ".pdf",
@@ -39,267 +28,6 @@ export const TEMPLATE_BRIEF_REFERENCE_ACCEPT = [
   "image/jpeg",
   "image/webp",
 ].join(",");
-
-const DOCX_MIME = "application/vnd.openxmlformats-officedocument.wordprocessingml.document";
-const PDF_MIME = "application/pdf";
-const TEMPLATE_BRIEF_AUTOFILL_TEXT_LIMIT = 12_000;
-const TEMPLATE_BRIEF_AUTOFILL_FIELD_LIMIT = 700;
-
-const TEMPLATE_BRIEF_REFERENCE_MIMES = new Set([
-  PDF_MIME,
-  DOCX_MIME,
-  "text/markdown",
-  "text/plain",
-  "text/csv",
-  "application/csv",
-  "application/json",
-  "image/png",
-  "image/jpeg",
-  "image/webp",
-]);
-
-const TEMPLATE_BRIEF_REFERENCE_MIME_BY_EXTENSION: Record<string, string> = {
-  pdf: PDF_MIME,
-  docx: DOCX_MIME,
-  md: "text/markdown",
-  txt: "text/plain",
-  png: "image/png",
-  jpg: "image/jpeg",
-  jpeg: "image/jpeg",
-  webp: "image/webp",
-  csv: "text/csv",
-  json: "application/json",
-};
-
-function templateBriefReferenceExtension(name: string) {
-  return name.split(".").pop()?.trim().toLowerCase() ?? "";
-}
-
-function templateBriefReferenceNameStem(name: string) {
-  const trimmed = name.trim();
-  const withoutPath = trimmed.split(/[/\\]/).pop() ?? trimmed;
-  const withoutExtension = withoutPath.replace(/\.[^.]+$/, "");
-  return withoutExtension.replace(/[_-]+/g, " ").replace(/\s+/g, " ").trim();
-}
-
-export const isTemplateBriefReferenceFile = isReferenceFile;
-export { inferTemplateBriefFromIngestions };
-
-export async function prepareTemplateBriefReferenceAttachment(file: File): Promise<ComposerAttachment> {
-  const isDocx = templateBriefReferenceExtension(file.name) === "docx" || templateBriefReferenceMime(file) === DOCX_MIME;
-  if (!isDocx) return prepareOriginalReferenceAttachment(file);
-
-  if (file.size > TEMPLATE_BRIEF_REFERENCE_MAX_BYTES) {
-    throw new Error(`${file.name} is larger than 25 MB.`);
-  }
-  if (!isTemplateBriefReferenceFile(file)) {
-    throw new Error(`${file.name} is not a supported reference document.`);
-  }
-
-  return {
-    id: `${file.name}-${file.lastModified}-${Math.random().toString(36).slice(2)}`,
-    name: file.name,
-    mimeType: "text/plain",
-    size: file.size,
-    kind: "file",
-    file: new File([await extractDocxText(file)], `${file.name}.txt`, { type: "text/plain" }),
-  };
-}
-
-function templateBriefReferenceMime(file: Pick<File, "name" | "type">) {
-  const mime = file.type.trim().toLowerCase();
-  if (TEMPLATE_BRIEF_REFERENCE_MIMES.has(mime)) return mime;
-  return TEMPLATE_BRIEF_REFERENCE_MIME_BY_EXTENSION[templateBriefReferenceExtension(file.name)] ?? "text/plain";
-}
-
-function decodeXmlText(value: string) {
-  return value
-    .replace(/&lt;/g, "<")
-    .replace(/&gt;/g, ">")
-    .replace(/&amp;/g, "&")
-    .replace(/&quot;/g, "\"")
-    .replace(/&apos;/g, "'");
-}
-
-async function extractDocxText(file: File) {
-  const { default: JSZip } = await import("jszip");
-  const zip = await JSZip.loadAsync(await file.arrayBuffer());
-  const documentXml = await zip.file("word/document.xml")?.async("string");
-  if (!documentXml) return "";
-  const paragraphs = documentXml
-    .split(/<\/w:p>/)
-    .map((paragraph) => [...paragraph.matchAll(/<w:t\b[^>]*>([\s\S]*?)<\/w:t>/g)]
-      .map((match) => decodeXmlText(match[1] ?? ""))
-      .join(""))
-    .map((paragraph) => paragraph.trim())
-    .filter(Boolean);
-  return paragraphs.join("\n");
-}
-
-function isTemplateBriefPdfReference(file: Pick<File, "name" | "type">) {
-  return templateBriefReferenceExtension(file.name) === "pdf" || templateBriefReferenceMime(file) === PDF_MIME;
-}
-
-function isTemplateBriefTextReference(file: Pick<File, "name" | "type">) {
-  const extension = templateBriefReferenceExtension(file.name);
-  if (extension === "md" || extension === "txt" || extension === "csv" || extension === "json") return true;
-  const mime = file.type.trim().toLowerCase();
-  return mime.startsWith("text/") || mime === "application/json" || mime === "application/csv";
-}
-
-function cleanTemplateBriefLine(value: string) {
-  return value
-    .replace(/^\s{0,3}#{1,6}\s*/, "")
-    .replace(/^\s*[-*+]\s+/, "")
-    .replace(/^\s*\d+[.)]\s+/, "")
-    .replace(/\*\*([^*]+)\*\*/g, "$1")
-    .replace(/__([^_]+)__/g, "$1")
-    .replace(/`([^`]+)`/g, "$1")
-    .trim();
-}
-
-function cleanTemplateBriefSnippet(value: string, limit = TEMPLATE_BRIEF_AUTOFILL_FIELD_LIMIT) {
-  return value
-    .split(/\r?\n/)
-    .map(cleanTemplateBriefLine)
-    .filter(Boolean)
-    .join(" ")
-    .replace(/\s+/g, " ")
-    .slice(0, limit)
-    .trim();
-}
-
-function firstMatchingLabelValue(text: string, labels: RegExp[]) {
-  const lines = text.split(/\r?\n/);
-  for (const line of lines) {
-    const cleaned = cleanTemplateBriefLine(line);
-    for (const label of labels) {
-      const match = cleaned.match(label);
-      const value = cleanTemplateBriefSnippet(match?.[1] ?? "", 240);
-      if (value) return value;
-    }
-  }
-  return "";
-}
-
-function markdownSectionText(text: string, keywords: string[]) {
-  const lines = text.split(/\r?\n/);
-  const sections: { heading: string; body: string[] }[] = [];
-  let current: { heading: string; body: string[] } | null = null;
-
-  for (const line of lines) {
-    const heading = line.match(/^\s{0,3}#{1,6}\s+(.+?)\s*#*\s*$/);
-    if (heading) {
-      current = { heading: cleanTemplateBriefLine(heading[1] ?? ""), body: [] };
-      sections.push(current);
-      continue;
-    }
-    current?.body.push(line);
-  }
-
-  const match = sections.find((section) => {
-    const heading = section.heading.toLowerCase();
-    return keywords.some((keyword) => heading.includes(keyword.toLowerCase()));
-  });
-  return cleanTemplateBriefSnippet(match?.body.join("\n") ?? "");
-}
-
-function firstBodyExcerpt(text: string, title: string) {
-  const lines = text
-    .split(/\r?\n/)
-    .map(cleanTemplateBriefLine)
-    .filter((line) => line && line !== title)
-    .filter((line) => !/^#{1,6}\s/.test(line));
-  return cleanTemplateBriefSnippet(lines.join("\n"));
-}
-
-function inferredTitleFromText(text: string, filename: string) {
-  const markdownTitle = text.match(/^\s{0,3}#\s+(.+?)\s*#*\s*$/m);
-  const cleanedMarkdownTitle = cleanTemplateBriefSnippet(markdownTitle?.[1] ?? "", 120);
-  if (cleanedMarkdownTitle) return cleanedMarkdownTitle;
-
-  const labelTitle = firstMatchingLabelValue(text, [
-    /^(?:title|presentation title|project name|name|标题|演示标题|作品标题|项目名称|名称)\s*[:：]\s*(.+)$/i,
-  ]);
-  if (labelTitle) return labelTitle;
-
-  const firstLine = text
-    .split(/\r?\n/)
-    .map(cleanTemplateBriefLine)
-    .find((line) => line.length > 1 && line.length <= 80 && !/[:：]\s*$/.test(line));
-  return firstLine ?? templateBriefReferenceNameStem(filename);
-}
-
-export function templateBriefFromReferenceText({ filename, text }: { filename: string; text: string }): TemplateBrief {
-  const source = text.replace(/^\uFEFF/, "").slice(0, TEMPLATE_BRIEF_AUTOFILL_TEXT_LIMIT);
-  const title = inferredTitleFromText(source, filename);
-  const audience = firstMatchingLabelValue(source, [
-    /^(?:audience|target audience|who it is for|users?|customers?|decision maker|受众|目标受众|面向谁|面向|目标用户|汇报对象|决策者|对象)\s*[:：]\s*(.+)$/i,
-    /^(?:purpose|goal|objective|decision|目标|目的|希望推动|希望推动什么决策|决策)\s*[:：]\s*(.+)$/i,
-  ]) || markdownSectionText(source, [
-    "audience",
-    "target audience",
-    "who it is for",
-    "decision",
-    "purpose",
-    "goal",
-    "objective",
-    "受众",
-    "目标受众",
-    "面向",
-    "目标用户",
-    "汇报对象",
-    "决策",
-    "目的",
-    "目标",
-    "希望推动",
-  ]);
-  const details = firstMatchingLabelValue(source, [
-    /^(?:details?|key information|requirements?|include|content|data|scope|信息|数据|内容|要求|包含|需要包含|核心内容|关键内容)\s*[:：]\s*(.+)$/i,
-  ]) || markdownSectionText(source, [
-    "key information",
-    "requirements",
-    "include",
-    "content",
-    "data",
-    "scope",
-    "background",
-    "overview",
-    "信息",
-    "数据",
-    "内容",
-    "要求",
-    "包含",
-    "核心内容",
-    "关键内容",
-    "背景",
-    "概述",
-    "方案",
-    "问题",
-  ]) || firstBodyExcerpt(source, title);
-
-  return {
-    title: title.trim(),
-    audience: cleanTemplateBriefSnippet(audience, 360),
-    details: cleanTemplateBriefSnippet(details),
-  };
-}
-
-export async function extractTemplateBriefReferenceText(file: File) {
-  const extension = templateBriefReferenceExtension(file.name);
-  const text = extension === "docx" || file.type.trim().toLowerCase() === DOCX_MIME
-    ? await extractDocxText(file)
-    : isTemplateBriefPdfReference(file)
-      ? ""
-    : isTemplateBriefTextReference(file)
-      ? await file.text()
-      : "";
-  return text.slice(0, TEMPLATE_BRIEF_AUTOFILL_TEXT_LIMIT);
-}
-
-export async function inferTemplateBriefFromReferenceFile(file: File): Promise<TemplateBrief> {
-  return inferTemplateBriefFromIngestions([await ingestReferenceFile(file)]);
-}
 
 export type TemplateBriefFields = TemplateBrief;
