+++ apps/app/tests/template-brief.test.ts
@@ -1,91 +1,16 @@
 import { beforeEach, describe, expect, test } from "bun:test";
 
 import { setLocale } from "../src/i18n";
-import JSZip from "jszip";
 
 import {
-  isTemplateBriefReferenceFile,
   isVideoStudioReady,
-  inferTemplateBriefFromReferenceFile,
-  prepareTemplateBriefReferenceAttachment,
-  templateBriefFromReferenceText,
   templateBriefConfigFor,
   templateBriefPrompt,
 } from "../src/react-app/domains/session/templates/template-brief";
-import {
-  buildTemplateReferenceSubmitPayload,
-  revokeTemplateReferenceAttachmentPreviews,
-} from "../src/react-app/domains/session/references/template-reference-submit";
-import type { TemplateReferenceItem } from "../src/react-app/domains/session/references/types";
 
 describe("template brief", () => {
   beforeEach(() => {
     setLocale("en");
-  });
-
-  test("template reference submit payload sends context by default and raw files only by opt in", async () => {
-    const file = new File(["Audience: enterprise teams"], "launch.txt", { type: "text/plain" });
-    const reference: TemplateReferenceItem = {
-      id: "ref_1",
-      file,
-      fileName: "launch.txt",
-      mimeType: "text/plain",
-      size: file.size,
-      status: "ready",
-      sendOriginal: false,
-      ingestion: {
-        id: "ref_1",
-        fileName: "launch.txt",
-        mimeType: "text/plain",
-        size: file.size,
-        sourceMode: "memory",
-        extractedText: "Audience: enterprise teams",
-        summary: "File: launch.txt",
-        chunks: [{ id: "launch.txt:chunk:1", source: "launch.txt", text: "Audience: enterprise teams", tokenEstimate: 4 }],
-        quality: "high",
-        warnings: [],
-      },
-    };
-
-    const defaultPayload = await buildTemplateReferenceSubmitPayload([reference]);
-    expect(defaultPayload.contextPack.promptText).toContain("launch.txt");
-    expect(defaultPayload.attachments).toEqual([]);
-
-    const optInPayload = await buildTemplateReferenceSubmitPayload([{ ...reference, sendOriginal: true }]);
-    expect(optInPayload.attachments).toHaveLength(1);
-    expect(optInPayload.attachments[0]?.name).toBe("launch.txt");
-  });
-
-  test("revokes only created template reference preview URLs", () => {
-    const revoked: string[] = [];
-    const originalRevokeObjectURL = URL.revokeObjectURL;
-    URL.revokeObjectURL = (url) => { revoked.push(url); };
-
-    try {
-      revokeTemplateReferenceAttachmentPreviews([
-        {
-          id: "image",
-          name: "visual.png",
-          mimeType: "image/png",
-          size: 1,
-          kind: "image",
-          file: new File(["x"], "visual.png", { type: "image/png" }),
-          previewUrl: "blob:template-image",
-        },
-        {
-          id: "text",
-          name: "notes.txt",
-          mimeType: "text/plain",
-          size: 1,
-          kind: "file",
-          file: new File(["x"], "notes.txt", { type: "text/plain" }),
-        },
-      ]);
-    } finally {
-      URL.revokeObjectURL = originalRevokeObjectURL;
-    }
-
-    expect(revoked).toEqual(["blob:template-image"]);
   });
 
   test("asks website creators for a site-specific brief", () => {
@@ -205,163 +130,4 @@ describe("template brief", () => {
     expect(prompt).toContain("do not rebuild it as a generic split hero");
   });
 
-  test("accepts supported optional reference document formats", () => {
-    const files = [
-      new File(["%PDF"], "brief.pdf", { type: "application/pdf" }),
-      new File(["copy"], "notes.md", { type: "text/markdown" }),
-      new File(["copy"], "notes.txt", { type: "text/plain" }),
-      new File(["image"], "visual.PNG", { type: "image/png" }),
-      new File(["image"], "photo.jpg", { type: "image/jpeg" }),
-      new File(["image"], "photo.jpeg", { type: "image/jpeg" }),
-      new File(["image"], "mock.webp", { type: "image/webp" }),
-      new File(["a,b"], "metrics.csv", { type: "text/csv" }),
-      new File(['{"a":1}'], "data.json", { type: "application/json" }),
-      new File([""], "word.docx", { type: "application/vnd.openxmlformats-officedocument.wordprocessingml.document" }),
-    ];
-
-    expect(files.map(isTemplateBriefReferenceFile)).toEqual(files.map(() => true));
-  });
-
-  test("rejects unsupported optional reference document formats", () => {
-    expect(isTemplateBriefReferenceFile(new File(["deck"], "old-deck.pptx", { type: "application/vnd.openxmlformats-officedocument.presentationml.presentation" }))).toBe(false);
-    expect(isTemplateBriefReferenceFile(new File(["sheet"], "budget.xlsx", { type: "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" }))).toBe(false);
-    expect(isTemplateBriefReferenceFile(new File(["svg"], "logo.svg", { type: "image/svg+xml" }))).toBe(false);
-  });
-
-  test("keeps legacy docx reference attachments as extracted text before sending", async () => {
-    const zip = new JSZip();
-    zip.file("word/document.xml", `<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-      <w:document xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main">
-        <w:body>
-          <w:p><w:r><w:t>Brand voice</w:t></w:r></w:p>
-          <w:p><w:r><w:t>Use concise clinical language.</w:t></w:r></w:p>
-        </w:body>
-      </w:document>`);
-    const buffer = await zip.generateAsync({ type: "arraybuffer" });
-    const attachment = await prepareTemplateBriefReferenceAttachment(
-      new File([buffer], "reference.docx", { type: "application/vnd.openxmlformats-officedocument.wordprocessingml.document" }),
-    );
-
-    expect(attachment.name).toBe("reference.docx");
-    expect(attachment.mimeType).toBe("text/plain");
-    expect(attachment.kind).toBe("file");
-    expect(attachment.file.name).toBe("reference.docx.txt");
-    expect(await attachment.file.text()).toContain("Use concise clinical language.");
-  });
-
-  test("keeps PDF references as PDF attachments until a real PDF text extractor is available", async () => {
-    const pdf = [
-      "%PDF-1.4",
-      "1 0 obj",
-      "<< /Type /Page >>",
-      "stream",
-      "BT",
-      "(Clinical Handoff) Tj",
-      "(Audience: Ward 7 nurses deciding handoff priorities.) Tj",
-      "(Include escalation path, risk flags, checklist, and shift-owner notes.) Tj",
-      "ET",
-      "endstream",
-      "endobj",
-      "%%EOF",
-    ].join("\n");
-
-    const attachment = await prepareTemplateBriefReferenceAttachment(
-      new File([pdf], "clinical-handoff.pdf", { type: "application/pdf" }),
-    );
-
-    expect(attachment.name).toBe("clinical-handoff.pdf");
-    expect(attachment.mimeType).toBe("application/pdf");
-    expect(attachment.kind).toBe("file");
-    expect(attachment.file.name).toBe("clinical-handoff.pdf");
-  });
-
-  test("keeps image-only PDF references as PDF attachments when text cannot be extracted", async () => {
-    const attachment = await prepareTemplateBriefReferenceAttachment(
-      new File(["%PDF-1.4\n/Type /XObject /Subtype /Image\n%%EOF"], "scan.pdf", { type: "application/pdf" }),
-    );
-
-    expect(attachment.name).toBe("scan.pdf");
-    expect(attachment.mimeType).toBe("application/pdf");
-    expect(attachment.file.name).toBe("scan.pdf");
-  });
-
-  test("keeps low-confidence PDF text as a PDF attachment instead of autofilling garbage", async () => {
-    const pdf = [
-      "%PDF-1.4",
-      "(zh)",
-      "(9□¾¨□9□ñ□3È□X5□E□¼□ib□÷~Ð□´|ôX-´D1#□□□cÔ□)",
-      "%%EOF",
-    ].join("\n");
-    const file = new File([pdf], "north-window-treatment.pdf", { type: "application/pdf" });
-
-    const attachment = await prepareTemplateBriefReferenceAttachment(file);
-    const brief = await inferTemplateBriefFromReferenceFile(file);
-
-    expect(attachment.mimeType).toBe("application/pdf");
-    expect(attachment.file.name).toBe("north-window-treatment.pdf");
-    expect(brief).toEqual({ title: "", audience: "", details: "" });
-    expect(brief.details).toBe("");
-  });
-
-  test("ignores PDF producer metadata instead of autofilling it", async () => {
-    const pdf = [
-      "%PDF-1.7",
-      "(Chromium)",
-      "(Skia/PDF m92 D:20241102090758+00'00' ÎU1 ÛUÔ=ué)",
-      "%%EOF",
-    ].join("\n");
-    const file = new File([pdf], "receipt_从双线叙事下的人物塑造看《牡丹亭》中的哲学思....pdf", { type: "application/pdf" });
-
-    const attachment = await prepareTemplateBriefReferenceAttachment(file);
-    const brief = await inferTemplateBriefFromReferenceFile(file);
-
-    expect(attachment.mimeType).toBe("application/pdf");
-    expect(attachment.file.name).toBe(file.name);
-    expect(brief).toEqual({ title: "", audience: "", details: "" });
-  });
-
-  test("infers brief fields from markdown reference text", () => {
-    const brief = templateBriefFromReferenceText({
-      filename: "ignored.md",
-      text: `# Clinical Handoff
-
-## Audience
-Ward 7 nurses deciding handoff priorities.
-
-## Key information
-Escalation path, risk flags, checklist, and shift-owner notes.`,
-    });
-
-    expect(brief.title).toBe("Clinical Handoff");
-    expect(brief.audience).toContain("Ward 7 nurses");
-    expect(brief.details).toContain("Escalation path");
-  });
-
-  test("infers brief fields from docx reference files", async () => {
-    const zip = new JSZip();
-    zip.file("word/document.xml", `<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-      <w:document xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main">
-        <w:body>
-          <w:p><w:r><w:t>Clinical Handoff</w:t></w:r></w:p>
-          <w:p><w:r><w:t>Audience: Ward 7 nurses deciding handoff priorities.</w:t></w:r></w:p>
-          <w:p><w:r><w:t>Include escalation path, risk flags, checklist, and shift-owner notes.</w:t></w:r></w:p>
-        </w:body>
-      </w:document>`);
-    const buffer = await zip.generateAsync({ type: "arraybuffer" });
-    const brief = await inferTemplateBriefFromReferenceFile(
-      new File([buffer], "clinical-handoff.docx", { type: "application/vnd.openxmlformats-officedocument.wordprocessingml.document" }),
-    );
-
-    expect(brief.title).toBe("Clinical Handoff");
-    expect(brief.audience).toContain("Ward 7 nurses");
-    expect(brief.details).toContain("escalation path");
-  });
-
-  test("uses a readable filename fallback when reference text cannot be extracted", () => {
-    const brief = templateBriefFromReferenceText({ filename: "clinical-handoff.pdf", text: "" });
-
-    expect(brief.title).toBe("clinical handoff");
-    expect(brief.audience).toBe("");
-    expect(brief.details).toBe("");
-  });
 });
